Starting with stats refactoring.
