+++ src/main/command_executor.cpp
@@ -53,7 +53,7 @@ CommandExecutor::CommandExecutor(Options& options)
       d_symman(new SymbolManager(d_solver.get())),
       d_smtEngine(d_solver->getSmtEngine()),
       d_options(options),
-      d_stats("driver"),
+      d_stats(),
       d_result()
 {
 }
