+++ src/theory/arrays/array_info.cpp
@@ -470,20 +470,20 @@ void ArrayInfo::mergeInfo(const TNode a, const TNode b){
 
 
       if(s!= 0) {
-        d_avgIndexListLength.addEntry(s);
+        d_avgIndexListLength << s;
         ++d_listsCount;
       }
       s = lista_st->size();
       d_maxList.maxAssign(s);
       if(s!= 0) {
-        d_avgStoresListLength.addEntry(s);
+        d_avgStoresListLength << s;
         ++d_listsCount;
       }
 
       s = lista_inst->size();
       d_maxList.maxAssign(s);
       if(s!=0) {
-        d_avgInStoresListLength.addEntry(s);
+        d_avgInStoresListLength << s;
         ++d_listsCount;
       }
