+++ src/util/statistics.h
@@ -35,8 +35,6 @@ class Stat;
 class CVC4_PUBLIC StatisticsBase {
 protected:
 
-  static std::string s_regDelim;
-
   /** A helper class for comparing two statistics */
   struct StatCmp {
     bool operator()(const Stat* s1, const Stat* s2) const;
@@ -44,8 +42,6 @@ protected:
 
   /** A type for a set of statistics */
   typedef std::set< Stat*, StatCmp > StatSet;
-
-  std::string d_prefix;
 
   /** The set of statistics in this object */
   StatSet d_stats;
@@ -77,9 +73,6 @@ public:
 
   /** An iterator type over a set of statistics. */
   typedef iterator const_iterator;
-
-  /** Set the output prefix for this set of statistics. */
-  virtual void setPrefix(const std::string& prefix);
 
   /** Flush all statistics to the given output stream. */
   void flushInformation(std::ostream& out) const;
